logique metier 90%

+++ service-creneaux_requetes/src/main/java/sid/org/service/RequeteServiceImpl.java
@@ -10,12 +10,18 @@ import org.springframework.stereotype.Service;
 import org.springframework.web.client.HttpStatusCodeException;
 
 import sid.org.api.ChatConnect;
+import sid.org.api.CompetenceApi;
+import sid.org.api.UserConnect;
 import sid.org.classe.Chat;
+import sid.org.classe.Competence;
 import sid.org.classe.Requete;
+import sid.org.classe.Users;
 import sid.org.dao.RequeteRepository;
 import sid.org.dto.ChatDto;
 import sid.org.dto.RequeteDto;
+import sid.org.exception.APiUSerAndCompetenceException;
 import sid.org.exception.EntityAlreadyExistException;
+import sid.org.exception.ForbiddenException;
 import sid.org.exception.ResultNotFoundException;
 
 @Service
@@ -24,15 +30,22 @@ public class RequeteServiceImpl implements RequeteService {
 	private RequeteRepository requeteRepository;
 	@Autowired
 	private ChatConnect chatConnect;
+	@Autowired
+	CompetenceApi competenceApi;
+	@Autowired
+	UserConnect userConnect;
 
 	@Override
-	public Requete createRequete(RequeteDto requeteDto) throws EntityAlreadyExistException {
+	public Requete createRequete(RequeteDto requeteDto)
+			throws EntityAlreadyExistException, APiUSerAndCompetenceException {
 
 		Optional<Requete> requete = requeteRepository.findByIdUserAndIdComp(requeteDto.getIdUser(),
 				requeteDto.getIdComp());
 		if (requete.isPresent()) {
 			throw new EntityAlreadyExistException("la requete existe deja");
 		}
+
+		Competence competence = competenceApi.getCompetence(requete.get().getIdComp());
 
 		Requete requete1 = new Requete();
 		requete1.setIdComp(requeteDto.getIdComp());
@@ -45,11 +58,14 @@ public class RequeteServiceImpl implements RequeteService {
 
 	}
 
-	public void validerUneRequete(Long id, Long idUser1) throws ResultNotFoundException, EntityAlreadyExistException {
+	@Override
+	public void validerUneRequete(Long id, Long idUser1) throws ResultNotFoundException, EntityAlreadyExistException,
+			HttpStatusCodeException, APiUSerAndCompetenceException {
 		Optional<Requete> requete = requeteRepository.findById(id);
 		if (!requete.isPresent()) {
 			throw new ResultNotFoundException("requete introuvable");
 		}
+
 		requete.get().setStatut("valide");
 		try {
 			Chat chat = chatConnect.getChat(requete.get().getIdUser(), idUser1);
@@ -63,31 +79,51 @@ public class RequeteServiceImpl implements RequeteService {
 	}
 
 	@Override
-	public Requete getRequete(Long id) throws ResultNotFoundException {
+	public Requete getRequete(Long id, Long idUser)
+			throws ResultNotFoundException, ForbiddenException, APiUSerAndCompetenceException {
 
 		Optional<Requete> requete = requeteRepository.findById(id);
 		if (!requete.isPresent()) {
 			throw new ResultNotFoundException("le requete est introuvable");
 		}
-		return requete.get();
+
+		Users user = userConnect.getUser(idUser);
+		Competence comp = competenceApi.getCompetence(requete.get().getIdComp());
+
+		if (requete.get().getIdUser() == idUser || comp.getUser().getCodeUtilisateur() == idUser) {
+			return requete.get();
+		} else {
+			throw new ForbiddenException("Vous n'avez pas accès a cette requete");
+		}
 
 	}
 
 	@Override
-	public Page<Requete> getRequetes(Long idUser) {
+	public Page<Requete> getRequetes(Long idUser) throws APiUSerAndCompetenceException {
 		Pageable pageable = PageRequest.of(0, 2);
+		userConnect.getUser(idUser);
 		Page<Requete> requete = requeteRepository.findByIdUser(idUser, pageable);
 		return requete;
 
 	}
 
 	@Override
-	public void deleteRequete(Long id) throws ResultNotFoundException {
+	public void deleteRequete(Long id, Long idUser)
+			throws ResultNotFoundException, APiUSerAndCompetenceException, ForbiddenException {
 		Optional<Requete> requete = requeteRepository.findById(id);
 		if (!requete.isPresent()) {
 			throw new ResultNotFoundException("le requete est introuvable");
 		}
-		requeteRepository.delete(requete.get());
+
+		Users user = userConnect.getUser(idUser);
+		Competence comp = competenceApi.getCompetence(requete.get().getIdComp());
+
+		if (requete.get().getIdUser() == idUser || comp.getUser().getCodeUtilisateur() == idUser) {
+			requeteRepository.delete(requete.get());
+		} else {
+			throw new ForbiddenException("Vous n'avez pas accès a cette requete");
+		}
+
 	}
 
 }
